Move nvim config to a folder
Instead of symlinking a single file, now link a whole folder to allow
more flexible customization for nvim :)

There was some auto formatting in this patch for the Rakefile!

+++ Rakefile
@@ -1,11 +1,11 @@
-require 'rake'
+require "rake"
 
-desc 'symlink all dot files into the home directory'
+desc "symlink all dot files into the home directory"
 task :default do
-  @ignored_files = ['Rakefile', 'README.md', 'nvim.init.lua']
+  @ignored_files = %w[Rakefile README.md nvim.init.lua]
 
   perform("Creating empty folders") { create_folders }
-  perform("Updating dotfiles")      { git_pull }
+  perform("Updating dotfiles") { git_pull }
   link_files
   link_nvim_config
 end
@@ -15,16 +15,14 @@ def git_pull
 end
 
 def link_files
-  Dir['*'].each { |file| link_file(file) unless @ignored_files.include?(file) }
+  Dir["*"].each { |file| link_file(file) unless @ignored_files.include?(file) }
 end
 
 def link_file(file, new_location = nil)
   home = `echo $HOME`.strip
   new_location ||= "#{file}"
 
-  if File.exist?("#{home}/.#{new_location}")
-    `rm "#{home}/.#{new_location}"`
-  end
+  `rm "#{home}/.#{new_location}"` if File.exist?("#{home}/.#{new_location}")
 
   perform "linking ~/.#{new_location}" do
     `ln -s "$PWD/#{file}" "#{home}/.#{new_location}"`
@@ -32,19 +30,18 @@ def link_file(file, new_location = nil)
 end
 
 def link_nvim_config
-
-  link_file('nvim.init.lua', 'config/nvim/init.lua')
+  link_file("nvim", "config/nvim")
 end
 
 def create_folders
-  `mkdir -p ~/.config/nvim`
+  `mkdir -p ~/.config`
   `mkdir -p ~/.vimbackup`
 end
 
 def perform(name, &block)
   print name
   yield
-  print ' ' * (50 - name.size)
+  print " " * (50 - name.size)
   print "DONE"
   puts
 end
